Refactor: fixBundleIdentifier -> iosProjectSetup
