+++ generators/app/tasks/appSetup/iosProjectSetup.js
